S11P12D109-298 feat: 티켓 생성 기능에서, 이미지 업로드 기능 추가

+++ Colorpl-Backend/src/main/java/com/colorpl/show/domain/schedule/ShowScheduleRepository.java
@@ -4,7 +4,7 @@ import com.colorpl.show.domain.detail.ShowDetail;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShowScheduleRepository extends JpaRepository<ShowSchedule, Integer> {
+public interface ShowScheduleRepository extends JpaRepository<ShowSchedule, Long> {
 
     List<ShowSchedule> findByShowDetail(ShowDetail showDetail);
 }
